requirements.txt
playwright
python-dotenv
cryptography